+++ scripts/macros.js
@@ -205,6 +205,8 @@ import {wrathOfTheStorm} from './macros/classFeatures/cleric/tempestDomain/wrath
 import {zealousPresence} from './macros/classFeatures/barbarian/zealot/zealousPresence.js';
 import {zombie} from './macros/monsterFeatures/zombie/zombie.js';
 import {zoneOfTruth} from './macros/spells/zoneOfTruth.js';
+import {kobold} from './macros/monsterFeatures/kobold/kobold.js';
+import {chromaticOrb} from './macros/spells/chromaticOrb.js';
 export async function onHitMacro(workflow) {
     if (workflow.targets.size === 0) return;
     workflow.targets.forEach(async token => {
@@ -261,7 +263,8 @@ let monster = {
     'succubus': succubus,
     'troglodyte': troglodyte,
     'undeadSpirit': undeadSpirit,
-    'zombie': zombie
+    'zombie': zombie,
+    'kobold': kobold
 }
 let onHit = {
     'blackrazor': blackrazor.onHit,
@@ -479,5 +482,6 @@ export let macros = {
     'witherAndBloom': witherAndBloom,
     'wrathOfTheStorm': wrathOfTheStorm,
     'zealousPresence': zealousPresence,
-    'zoneOfTruth': zoneOfTruth
+    'zoneOfTruth': zoneOfTruth,
+    'chromaticOrb': chromaticOrb
 }