+++ scripts/macros/monsterFeatures/kobold/kobold.js
@@ -1,0 +1,4 @@
+import {pactTactics} from './packTactics.js';
+export let kobold = {
+    'packTactics': pactTactics
+}